Move class to dedicated file
