+++ vector/src/main/java/im/vector/riotx/features/home/createdirect/CreateDirectRoomActions.kt
@@ -19,7 +19,6 @@ package im.vector.riotx.features.home.createdirect
 import im.vector.matrix.android.api.session.user.model.User
 
 sealed class CreateDirectRoomActions {
-
     object CreateRoomAndInviteSelectedUsers : CreateDirectRoomActions()
     data class FilterKnownUsers(val value: String) : CreateDirectRoomActions()
     data class SearchDirectoryUsers(val value: String) : CreateDirectRoomActions()
